+++ src/Deformula.jl
@@ -1,5 +1,0 @@
-module Deformula
-
-include("_Deformula.jl")
-
-end