+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.0
+
+- Change default value of maxiter from 16 to 12
+- Update for BigFloat support
+- Change the name of the package from Deformula to DEQuadrature
+
 # v0.1.4
 
 - Update compatibility to Julia 1.6 and above.